Memoize convertColor to add speed.
Also, refactor util/index.js to have color utilities in their own file
and other common utilities in util/common.js.

Named CSS colors are now taken from the npm color-name package instead
of including them ourselves.
